app/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.3
django-admin==2.0.2
django-cors-headers==3.11.0
django-excel-response2==3.0.6
django-filter==21.1
django-grappelli==3.0.3
django-six==1.0.5
django-templated-mail==1.1.1
django-tinymce==4.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
excel-base==1.0.4
idna==3.3
importlib_metadata==7.1.0
isoweek==1.3.3
Markdown==3.6
oauthlib==3.2.0
pillow==10.3.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
screen==1.0.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
TimeConvert==3.0.13
typing_extensions==4.10.0
tzlocal==5.2
urllib3==1.26.8
xlwt==1.3.0
zipp==3.18.1
gunicorn==21.2.0